Refactor: Migrate main content to Nuxt page component
- Moved entire app content from app.vue to pages/index.vue
- Simplified app.vue to use NuxtLayout and NuxtPage
- Preserved all existing functionality and styling
- Improved project structure for Nuxt 3 routing
